requirements.txt
python-dotenv==0.21.1
pandas==1.5.3
